+++ test/install_generator_test.rb
@@ -13,8 +13,7 @@ class InstallGeneratorTest < Rails::Generators::TestCase
     run_generator
 
     assert_migration("db/migrate/install_online_migrations.rb") do |content|
-      assert_includes content, "create_table :background_migrations"
-      assert_includes content, "create_table :background_migration_jobs"
+      assert_includes content, "create_table :background_data_migrations"
       assert_includes content, "create_table :background_schema_migrations"
     end
   end