+++ test/schema_statements/indexes_test.rb
@@ -113,7 +113,9 @@ module SchemaStatements
     end
 
     def test_add_index_in_background
-      m = @connection.add_index_in_background(:users, :name, unique: true, connection_class_name: "User")
+      @connection.add_index_in_background(:users, :name, unique: true, connection_class_name: "User")
+      m = last_schema_migration
+
       assert_equal "index_users_on_name", m.name
       assert_equal "users", m.table_name
       assert_equal 'CREATE UNIQUE INDEX CONCURRENTLY "index_users_on_name" ON "users" ("name")', m.definition
@@ -174,7 +176,9 @@ module SchemaStatements
     end
 
     def test_add_index_in_background_custom_attributes
-      m = @connection.add_index_in_background(:users, :name, name: "my_name", max_attempts: 5, statement_timeout: 10, connection_class_name: "User")
+      @connection.add_index_in_background(:users, :name, name: "my_name", max_attempts: 5, statement_timeout: 10, connection_class_name: "User")
+      m = last_schema_migration
+
       assert_equal "my_name", m.name
       assert_equal 5, m.max_attempts
       assert_equal 10, m.statement_timeout
@@ -206,7 +210,9 @@ module SchemaStatements
 
     def test_remove_index_in_background_custom_attributes
       @connection.add_index(:users, :name)
-      m = @connection.remove_index_in_background(:users, :name, name: "index_users_on_name", max_attempts: 5, statement_timeout: 10, connection_class_name: "User")
+      @connection.remove_index_in_background(:users, :name, name: "index_users_on_name", max_attempts: 5, statement_timeout: 10, connection_class_name: "User")
+      m = last_schema_migration
+
       assert_equal "index_users_on_name", m.name
       assert_equal 5, m.max_attempts
       assert_equal 10, m.statement_timeout