+++ test/support/schema.rb
@@ -5,64 +5,35 @@ ActiveRecord::Schema.define do
   enable_extension "citext"
   enable_extension "btree_gist"
 
-  create_table :background_migrations, force: :cascade do |t|
-    t.bigint :parent_id
+  create_table :background_data_migrations do |t|
     t.string :migration_name, null: false
     t.jsonb :arguments, default: [], null: false
-    t.string :batch_column_name, null: false
-    t.bigint :min_value, null: false
-    t.bigint :max_value, null: false
-    t.bigint :rows_count
-    t.integer :batch_size, null: false
-    t.integer :sub_batch_size, null: false
-    t.integer :batch_pause, null: false
-    t.integer :sub_batch_pause_ms, null: false
-    t.integer :batch_max_attempts, null: false
     t.string :status, default: "enqueued", null: false
     t.string :shard
-    t.boolean :composite, default: false, null: false
+    t.string :cursor
+    t.string :jid
     t.datetime :started_at
     t.datetime :finished_at
-    t.timestamps
-
-    t.foreign_key :background_migrations, column: :parent_id, on_delete: :cascade
-
-    t.index [:migration_name, :arguments, :shard],
-      unique: true, name: :index_background_migrations_on_unique_configuration
-  end
-
-  create_table :background_migration_jobs, force: :cascade do |t|
-    t.bigint :migration_id, null: false
-    t.bigint :min_value, null: false
-    t.bigint :max_value, null: false
-    t.integer :batch_size, null: false
-    t.integer :sub_batch_size, null: false
-    t.integer :pause_ms, null: false
-    t.datetime :started_at
-    t.datetime :finished_at
-    t.string :status, default: "enqueued", null: false
+    t.bigint :tick_total
+    t.bigint :tick_count, default: 0, null: false
+    t.float :time_running, default: 0.0, null: false
     t.integer :max_attempts, null: false
-    t.integer :attempts, default: 0, null: false
     t.string :error_class
     t.string :error_message
     t.string :backtrace, array: true
+    t.string :connection_class_name
     t.timestamps
 
-    t.foreign_key :background_migrations, column: :migration_id, on_delete: :cascade
-
-    t.index [:migration_id, :max_value], name: :index_background_migration_jobs_on_max_value
-    t.index [:migration_id, :status, :updated_at], name: :index_background_migration_jobs_on_updated_at
-    t.index [:migration_id, :finished_at], name: :index_background_migration_jobs_on_finished_at
+    t.index [:migration_name, :arguments, :shard],
+      unique: true, name: :index_background_data_migrations_on_unique_configuration
   end
 
-  create_table :background_schema_migrations, force: :cascade do |t|
-    t.bigint :parent_id
+  create_table :background_schema_migrations do |t|
     t.string :migration_name, null: false
     t.string :table_name, null: false
     t.string :definition, null: false
     t.string :status, default: "enqueued", null: false
     t.string :shard
-    t.boolean :composite, default: false, null: false
     t.integer :statement_timeout
     t.datetime :started_at
     t.datetime :finished_at
@@ -74,8 +45,7 @@ ActiveRecord::Schema.define do
     t.string :connection_class_name
     t.timestamps
 
-    t.foreign_key :background_schema_migrations, column: :parent_id, on_delete: :cascade
-
-    t.index [:migration_name, :shard], unique: true, name: :index_background_schema_migrations_on_unique_configuration
+    t.index [:migration_name, :shard, :connection_class_name], unique: true,
+      name: :index_background_schema_migrations_on_unique_configuration
   end
 end